高分险算分逻辑

+++ ruoyi-admin/src/main/java/com/ruoyi/electrical/dto/IDangerExportDto.java
@@ -14,5 +14,5 @@ public interface IDangerExportDto {
 
 	void setOpenStatus(String openStatus);
 
-	void setTotalScore(Long totalScore);
+	void setTotalScore(Double totalScore);
 }
